rules: Infallible diagnostics in extra outputs
Summary: Tbh, I'm not really sure yet what the right thing to put into this provider is. The previous version had both the metadata outputs, but not the infallible diagnostics available. I don't know if anything is ever going to have a use case for that, but for now I'm adjusting this type to serve the use case that I know exists - rust analyzer check builds. We can revisit more of this API later

Reviewed By: capickett

Differential Revision: D59886990

fbshipit-source-id: daa6d7ce973af98a82b7712327bac1999e441918