Fix handling of multing constraints for same setting
Summary:
Previously, if a user passed different constraint values on the same
setting, we'd get somewhat unintuitive results, as we'd filter out ones
that were already set on the platform.

This fixes by doing a pass on the incoming constraints to take only
the last one for a particular setting.

Reviewed By: JakobDegen

Differential Revision: D60019447

fbshipit-source-id: 1487a1ac2c3e9fbee9ae3a6efbc3e5a7fdb58865