Stop attempting to give "native" archives special treatment
Summary:
D47336408 attempts to circumvent some operations for when an archive contains
only native object files instead of bitcode. I suspect we get almost no benefit
out of this, as only pre-built c++ libraries are marked as "native":

https://www.internalfb.com/code/fbsource/[91f7067b80f8]/fbcode/buck2/prelude/cxx/prebuilt_cxx_library_group.bzl?lines=133

More importantly, this complicates the system and re-introduces a bug which D38925474
fixed.

Avoiding extra work for native input files might not be a bad idea, but I don't think
this is the way to go about it. We can revist later.

This diff effectively reverts the parts of D47336408 that are in the Darwin fork of
the dist_lto logic.

Reviewed By: rmaz

Differential Revision: D60181464

fbshipit-source-id: de9ba57a9ee51569965d4714318664973fb51747