Remove [linker.command] subtarget information for gnu dthin-lto links
Summary:
In D59840767 I modified the GNU dthin-lto logic to produce two sets of flags returned
in the produced LinkedObject.

These flags are analogous to the existing index_argsfile,
and linker_argsfile fields, they describe thin-link and codegen flags instead
of thin-link and native-link flags, and do not include inputs. That is the names
of the object files passed to the linker are not included.

This is because we wanted this information to be used in tests comparing
various sets of flags for particular builds. This information needs to be fetched
quickly, we can't depend upon compilation of all dependencies. For dthin-lto,
we need to inspect input archives to determine the shape of the linker invocation,
so we just leave out all inputs all together.

The Darwin version of this file has refactored this new logic in D60255806 and D60181464,
to fix missing flags. That means as is, this set of flags is incomplete.

GNU builds don't use this information, and unless someone thinks it will be particularly
useful, I would rather remove it. This way it doesn't remain unused and unmaintained.

Reviewed By: milend

Differential Revision: D60307970

fbshipit-source-id: e0789f5e442b268a3989b7c6d2ab4b684c6a722d