Provide DWP for shared libs as additional outputs when requesting DWP for top level targets
Summary:
Please see the discussion in https://fb.workplace.com/groups/linker.support/posts/1228209065012600 for more context.

Users may find it surprising that requesting `[dwp]` for the top level target doesn't also generate DWP associated with shared library dependencies. This diff provides those as as additional outputs when building cxx and rust executables so that users don't have to manually find and build the `[dwp]` subtarget associated with each shared library.

If the existing behavior is preferable or if there are better ways to provide this functionality, please leave a comment on the diff :)

Reviewed By: luciang

Differential Revision: D60475856

fbshipit-source-id: c140e388acdaec15b2e7f8e02be9ff395461430b