r-a: Re-enable incremental for clippy
Summary:
While fixing the bug.

This has to be up there for the dumbest bugs I ever wrote.

I feel ok not writing an automated test in this diff (although I should) because I'm working on some stuff that'll improve the testing situation for rust rules in the long term

Reviewed By: Wilfred, davidbarsky, dtolnay

Differential Revision: D60608655

fbshipit-source-id: 0e3e05872bc157b38c92ecde4d8aaec0a0dda4a8