buck2 | xcode | Expose apple_resources to Xcode
Summary:
Projects with .xib files and other apple_resources can't currently see / edit them from the Xcode project.

The build works, because it uses Buck. All we need is to add entries for these files in the Buck project to expose them.

This diff attempts this by populating the "extra_xcode_files" attribute for this target. ("other_files" did not work for some reason).

Differential Revision: D60167547

fbshipit-source-id: 47edcb1a6b7845f82a14ac2361ae23d7dd7d8aee