Only add coverage instrumentation for clang
Summary:
`-fprofile-instr-generate` and `-fcoverage-mapping` are only available on clang.

In D60771652, there are cuda targets (which are built using gcc, even in dev/dev-nosan mode) that depend on NetworkFlow.h, which is one of the selective coverage targets. The cuda files are added into the set of tests that need coverage and then these flags (`-fcoverage-mapping`, ...) are added, and then the compilation fails because these are invalid gcc flags.

So in this diff: if you encounter instrumentation flags, check the compiler type and only add the flags if the compiler type is clang.

Reviewed By: kit1980, aniketmathur, jbardini

Differential Revision: D61219425

fbshipit-source-id: f8e5d77eab6731158436836a4b28c88c3e178387