Enabling test rules to run on Mac RE
Summary: When migrating jobs from mac to Linux, builds can happen on Linux machines but tests can't. To get mac specific builds and test only executing on a mac host, the exec platform would need to be constrained to prefer a mac host. However that causes both builds (which can be done on a linux system) and tests to execute on a mac, and this doesn't efficiently utilize the mac fleet

Reviewed By: peterdelevoryas

Differential Revision: D60912331

fbshipit-source-id: 54c978d539f964e78f67a61a7cc71e1bcc43a07d