rules: Bring back `rustc_check_flags`
Summary:
This was deleted at some point because it didn't make sense (D55438348). And for the old interpretation of the flag, that was indeed correct, since it allowed for divergence between metadata-full and link actions, which should not be allowed and cause "can't find crate errors."

This new version prevents that by only applying this flag to the metadata-fast graph, which is separate from the metadata-full/link graph and so doesn't have that problem.

The use case is in the next diff

Reviewed By: dtolnay

Differential Revision: D61447188

fbshipit-source-id: 3f2f4589b721698181653ed425286813d99741e3