support windows shared library exports via deffiles
Summary:
Windows has two ways of specifying what functions to export from a shared library (dll). You can either declare the functions with __declspec(dllexport), or you can provide that and more options in a *.def file.

Today, the only way to use a def file is to setup the linker arguments in your `cpp_library` build rule yourself with a select-on-windows statement. Given the support already in place for setting `soname`, the import library, etc, this adds a `deffile` argument to the `cpp_library` and friends, which adds the appropriate linker command line argument.

Reviewed By: akrieger

Differential Revision: D49229817

fbshipit-source-id: 0473b0fce003da3e70e7e8565615f4335ee85318