Enabled link groups for shared linkage in rust
Summary:
# What we are doing
We implementing link groups support for dynamic linking. Current problem is: with following setup
```
    :bin
    /   \
 :A      :C
    \   /
      :B
```

if we put `:A` and `:B` to link group, `lib_c.so` will still add `lib_b.so` to `NEEDS` section. But `lib_b.so` is gonna be grouped to `lib_a_b_lg.so` and there is no way in BUCK to propagate this information to `lib_c.so`.
The good part is we already build link tree on file system via symlinks. So we actually can have "stubs" for `lib_a.so` and `lib_b.so` that all point to actual `lib_a_b_lg.so`. Dynamic linked is perfectly fine with that.

## What this diff is doing
That is the same change as in previous diff but for link groups in rust

For each shared library (that derived from library target) we determine if it actually went to any link group. If it is we create additional symlink that will point to link group. It will look like:
```
$ tree buck_out/__bin_symlink_tree__
__bin_symlink_tree__
├ lib_a_b_lg.so -> ../../__link_groups__/lib_a_b_lg.so
├ lib_a.so -> ../../__link_groups__/lib_a_b_lg.so
└ lib_b.so -> ../../__link_groups__/lib_a_b_lg.so
```

Reviewed By: capickett

Differential Revision: D58787726

fbshipit-source-id: a82f74bb6377b76d155d2413563a4e5fb408b4e7