Use a separate package db when gathering package metadata
Using multiple source directories results in multiple entries in the `import-dirs` field of the package conf file.

With multiple entries in that field, GHC tries to find all interface files for the exposed modules (see [1] and [2]).

For the dependency analysis we do not want to depend on any interface files. To make this work we have to ensure there only is a single entry. But also, the multiple entries are needed for compilation (using the empty package dbs). That is why yet another package db is introduced here.

[1]: https://gitlab.haskell.org/ghc/ghc/-/blob/3a145315052d6f66f9682ecff87b522011165d59/compiler/GHC/Unit/Finder.hs#L491-499
[2]: https://github.com/tweag/mercury-ghc-internal/issues/51