rust-project: Add flag to disable workspaces
Summary: Fixing `with_rust_workspace` will cause a slight regression in the glean indexer since targets in the same workspace shuffled to different shards will result in duplicate work (both shards will index the entire workspace). This diff adds a flag to disable the behavior that we can use in the indexer.

Reviewed By: Wilfred

Differential Revision: D61605515

fbshipit-source-id: e935347ad0ba07d202fead49756cbee0ecbc4168