rust-project: Properly limit extra targets
Summary:
In D61174858, I mentioned
> Technically this behavior is still not quite right. It is now "the max number of extra targets per buildfile", but that's fine.

Since the glean indexer is the only client of `max_extra_targets`, it really is fine, but in trying to fix `with_rust_workspace`, I discovered that we implicitly depend on including extra targets to ensure we get both `foo` and `foo-unittest` from `resolve_owning_buildfile` (see D61543769 for more context). That is, if `max_extra_targets=0`, we will *arbitrarily* get *either* `foo` or `foo-unittest` when querying the owner of `foo.rs`.

Again, this is *probably* still fine since rust-project uses `max_extra_targets=50`, but it's technically possible to fail, makes testing harder, and made me just uncomfortable enough to fix it properly.

Reviewed By: Wilfred

Differential Revision: D61559933

fbshipit-source-id: 418df511c72bf3ce542b5b64a8d8dfdcfb2448a2