Re-land no-rust exports change
Summary:
Fix and undo D61640971 .

What happens here is that the double level cxx usage and particular shape of the dep graph eventually ends up with a situation where someone depends directly (as a result of exporting) on `libcommon_rust_shed_stats_facebook_stats_facebook-bridge.so`, but not on the Rust library implementing some of the symbols that requires. When that happens, the linker sees when building an executable that some of the symbol references in the above dylib are unfulfilled, and so errors out. But there isn't actually a problem, we do provide implementations for those symbols, we just don't materialize the dylibs providing those implementations in that particular link because they're not needed there.

Fortunately, this is the linker equivalent of a lint, so we can just suppress it.

Reviewed By: capickett

Differential Revision: D61646321

fbshipit-source-id: 82559178b930ed3151de8ff55ff6bf6378420b0c