Do not create opt actions for unloaded bitcode files
Summary:
Distributed thin-lto thin-link takes all the input files and writes several artifacts to disk. These artifacts are read by a wrapper python script called the "dist_lto_planner". It reads the artifacts, and writes a series of JSON files called "plans".

These plan files are read by starlark as dynamic outputs, and used to inform the opt + codegen actions that need to be created.

One of the artifacts produced by distributed thin-lto thin-link is a "index file". This is the file pointed at by `-thinlto-index-only=` and it is the list of bitcode files parsed (loaded) by the linker.

Prior to this patch, we created opt + codegen actions for all input bitcode files, even though some do no participate in the link. Especially for apps like Wilde where we don't use `-ObjC`, this can be quite many unnecessary actions.

In addition, if we plan to move forward with a single unified index (by index I now mean serialized combined summary), we require this change. The serialized index only includes information for loaded input bitcode files. If we attempt to run a opt + codegen action for an unloaded bitcode file against a summary which does not include GV summaries for the GVs defined in that bitcode file, we will at least fail assertions.

Reviewed By: rmaz

Differential Revision: D62139842

fbshipit-source-id: a5bfd495e9a5f590d3f257ebced484c8fb39b36a