Remove execution_logs logic from test_runner.
Summary:
This logic was meta specific and should be left to a buck_ct_provider.

Tpx will scan execution for a footer.txt, in which additional data can be displayed to the user. It should be up to the user to provide a provider that will generate this additional footer.

Reviewed By: TheGeorge, Balajiganapathi

Differential Revision: D49912997

fbshipit-source-id: 18d3617049d5c3f099dc568742d13b9a48d2e668