python: No cxx toolchain in fat platforms
Summary:
Python rules support being built directly against fat platform configurations when they do not depend on any C++. However, currently when this happens the rules end up with access to a C++ toolchain that makes no sense for the actual target configuration. This diff fixes that.

In order to make this work, we have to error out in the python rules in the cases where the C++ toolchain would have been needed even without any C++ deps, specifically for platform attrs. It's possible that this will introduce failrues into some existing code, however to the extent that it does this is a bugfix. If it does turn out to be a problem, we can deal with that by adjusting macros to mark python targets as fat platform incompatible if they have non-empty `platform_` attrs

My concrete reason for wanting this is that I'm introducing more selects onto the C++ toolchain in D62157852 , and those selects break in fat target platforms.

Reviewed By: milend

Differential Revision: D62346759

fbshipit-source-id: 4ce4cfbe84342e6af10c8d1a18c17f5bf325405a