Fix erlang_test() not honoring the _toolchain attr when creating the test node
Summary:
# Context
The `erlang_test()` rule has an optional `_toolchain` attribute that can be used to override the default Erlang toolchain. That means that the sources are built with the given toolchain and, one expects, they are executed using that toolchain as well.

# Problem
At some point, some change was introduced that instead of using the `erl` binary from the given toolchain, whatever is in the path is used instead, which is incorrect. Afaiu, the motivation for this was to mitigate a slowness on platforms that rely on dotslash for toolchain distribution:

  - dotslash stores its cache of downloaded binaries in ~/.cache/dotslash
  - when running a test, the HOME variable is set to a temp directory, but this means that ~/.cache/dotslash doesn't exist, which will make dotslash re-download the toolchain on every test run

# This diff
We revert back to the correct behavior of using the `erl` that comes in the given toolchain. To mitigate the dotslash issues, we symlink `.cache/dotslash` in the temp HOME to the real dotslash cache, so that already downloaded artifacts can be reused. This doesn't impact hermeticity, since it is cache and file writing happens at test execution time.

Reviewed By: TheGeorge, robertoaloi

Differential Revision: D62196849

fbshipit-source-id: 95b1f49d1717c76e6d67b10d49e0bc416bbb9a1c