Create incremental_state subtarget
Summary:
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This stack
This stack focuses on e2e tests implementation.

### This diff
Based on the suggestion from the [post](https://fb.workplace.com/groups/buck2users/permalink/3726629097593444/), this diff creates an `incremental_state` subtarget to be able to expose the dir in e2e tests.

Reviewed By: IanChilds

Differential Revision: D62576747

fbshipit-source-id: 0cb1165474274a18439a15705cec61e97d3147bc