Fix tests not running on macos
Summary:
# Context
We recently added a feature that symlinks the real dotslash cache to the fake home directory the test-runner uses. While we tried to make it portable, unfortunately, a bug make it work by chance on Linux, and fail on Macos

# Problem
We need to ensure that the "cache" directory that contains the dotslash cache exists in the fake home directory where the symlink is created. However we were doing

```
filelib:ensure_dir(filename:dirname(FakeHomeDotslashCache)
```

but this is dropping the final "dotslash" twice, since `filelib:ensure_dir()` already ignores the final element of the given path. How was this even working on Linux? Whlie making it portable, we left a stray `mkdir(".cache")`

# This diff
Fix by using `ensure_path(dirname(...))` to make the intention more explicit

Reviewed By: zerth

Differential Revision: D62603643

fbshipit-source-id: 85dbe6b84e806be6951d760b07048054a5c382dc