Enable local test execution for iOS
Summary: Currently for iOS, `buck test` will always execute tests in RE. This diff adds options to force local or remote test execution, and adds a new automatic option to select where tests execute based on if local execution is possible. (ie. The user is on a Mac with Xcode installed.)

Reviewed By: jbardini

Differential Revision: D62206858

fbshipit-source-id: 07bf50ec3fb7879489d3e887ce49a25ae0f55b0d