Allow passing default_roots to rust_test to disambiguate
Summary:
In the general case, it's ambiguous whether a usage of `rust_test()` is to declare a test target for a *library* or for a *binary*. This makes it tricky to determine the crate root as we must consider both `lib.rs` and `main.rs`. However, for our use-case, we tend to add unit test targets as part of a `rust_library` or `rust_binary` macro, where we have the information to disambiguate.

This change proposes exposing `default_roots` on `rust_test` so that we can pass the correct root candidate for our automatic test targets.

See also https://fb.workplace.com/groups/rustonmobile/posts/3858526464393667

Reviewed By: dtolnay

Differential Revision: D62585839

fbshipit-source-id: d00b890d7b051e92ce4d16e388fe520ef05a8248