Use absolute paths for the argsfile containing argsfiles
Summary:
In D62179090, we made a change to break up the argsfile to direct to partial argsfiles. This was to allow for caching and reuse of partial argsfiles.

However, when Xcode indexes, it cannot find these partial files.

However, converting them to absolute paths does resolve the issue. So when we write the wrapping argsfile, let's ***use absolute paths for the partial argsfiles***.

Reviewed By: milend

Differential Revision: D63464898

fbshipit-source-id: 88f1053840c51e1776f2f6142e6798d029a3fe1c