Pull shared type check rule resources out into main type check definition function
Summary:
This diff moves several shared pieces for type checking into the `create_per_target_type_check` function. The signature of
`_create_batched_type_check` is changed to reflect the new data used. The next diff creates the implementation for defining each shard.
 ---

I'm going to be making type checking operations shardable, which should speed up type checking (especially for IDE use cases). The method of doing this will be to create sub-targets under the `typecheck` sub-target, all of which we will send to RE to calculate type errors.

The idea here is to kick off separate RE jobs for each file, where we type check only that file. All other files in the target (and all dependencies) are treated as dependencies, and are not type checked.

Followup: type check only the shards relating to the specific files that are open (for IDE use).

effective_sharding

Reviewed By: grievejia

Differential Revision: D64063896

fbshipit-source-id: 004d4e0727335b44ed0122bb240d3fe9fd29dad7